+++ lib/src/commands/commands.dart
@@ -1,2 +1,3 @@
 export 'init_command.dart';
 export 'make_command.dart';
+export 'new_command.dart';